🚸 (run) make outdir flag dependent on output flag
At the moment, it is technically possible to call the command with the outdir flag but not the ouput flag, which is basically a noop.

The output flag is the one telling the cli to write the files. It doesn't make sense to use outdir if you don't want to write the files. As a temporary improvement, make the outdir flag dependant on the output flag.

In the future we might want to merge the two flags.